refactor(settings): refactor ai-section components
file per component makes it reduces cognitive weight
